+++ source/classes/thermosphere/thermosphere.cpp
@@ -17,13 +17,15 @@
 #include "external/hwm/hwm.h"
 #include "config/configRegister.h"
 #include "classes/thermosphere/thermosphereJB2008.h"
+#include "classes/thermosphere/thermosphereNRLMSIS2.h"
 #include "classes/thermosphere/thermosphere.h"
 
 
 /***********************************************/
 
 GROOPS_REGISTER_CLASS(Thermosphere, "thermosphereType",
-                      ThermosphereJB2008)
+                      ThermosphereJB2008,
+                      ThermosphereNRLMSIS2)
 
 GROOPS_READCONFIG_CLASS(Thermosphere, "thermosphereType")
 
@@ -39,6 +41,8 @@ ThermospherePtr Thermosphere::create(Config &config, const std::string &name)
     readConfigChoice(config, name, choice, Config::MUSTSET, "", "density, temperature and velocity");
     if(readConfigChoiceElement(config, "jb2008",  choice, "Jacchia-Bowman 2008 Empirical Thermospheric Density Model"))
       thermosphere = ThermospherePtr(new ThermosphereJB2008(config));
+    if(readConfigChoiceElement(config, "nrlmsis2",  choice, "NRLMSIS 2.0 Empirical Thermospheric Density Model"))
+      thermosphere = ThermospherePtr(new ThermosphereNRLMSIS2(config));
     endChoice(config);
 
     return thermosphere;